+++ test/test_constraint.py
@@ -197,7 +197,7 @@ def test_constraint_wrapped_methods(x, y):
     con.rename({"first": "new_labels"})
     con.rename_dims({"first": "new_labels"})
     con.roll({"first": 1})
-    con.stack(new_dim=("first", "second"))
+    con.stack(new_dim=("first", "second")).unstack("new_dim")
 
 
 def test_anonymous_constraint_sel(x, y):